+++ spec/onlyoffice_rspec_result_parser/result_parser_get_total_result_of_rspec_html_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe OnlyofficeRspecResultParser::ResultParser,
+         '.get_total_result_of_rspec_html' do
+  it 'get_total_result_of_rspec_html return empty string for empty string' do
+    expect(described_class.get_total_result_of_rspec_html('')).to eq('')
+  end
+end